.codeium/windsurf/code_tracker/active/no_repo/7e224eee85ae61e590d7cf949df832ef_requirements.txt
ΩFlask==2.3.3
torch==2.9.1
torchvision==0.24.1
opencv-python==4.10.0.84
Pillow==10.4.0
numpy>=1.26.0
scipy>=1.13.0
matplotlib>=3.8.0
tqdm>=4.66.0
werkzeug==2.3.7
requests>=2.31.0
 *cascade08*cascade08, *cascade08,/*cascade08/B *cascade08BD*cascade08DE *cascade08EF*cascade08FH *cascade08HI*cascade08IU *cascade08UW*cascade08W` *cascade08`a*cascade08ae *cascade08ef*cascade08fg *cascade08gh*cascade08ho *cascade08op*cascade08pu *cascade08uw*cascade08wÉ *cascade08ÉÑ*cascade08Ñá *cascade08áà*cascade08àâ *cascade08âä*cascade08äê *cascade08êë*cascade08ëó *cascade08óò*cascade08ò≥ *cascade08≥¥*cascade08¥Ω *cascade0827file:///c:/Users/getan/Documents/final/requirements.txt